hotfix for consumer_consume errors causing potential hangs
3c7da1f7 added EOF handling for consumer threads which makes the
consumer thread exit without cleanup on error conditions. Given an
error condition on consumer threads, producer threads will hang forever
waiting for input from consumer threads which are no longer existent
causing a stall.

Previously consumer threads would not exit on error conditions and
continue to (attempt to) consume indefinitely which would have prevented
this issue in case the error condition is transient.

This commit is meant as a hotfix which ensures cleanup happens on
consumer errors which should work fine with a single consumer thread
(which is how we run schaufel on baggers at the moment). Proper fix to
handle more than one consumer thread is WIP.