file_consumer_consume: free buffer on EOF
getline(3) states:
```
This buffer should be freed by the user program even if getline()
failed.
```

The statement is partially misleading since EOF is signalled via -1 as
well so we do want to check errno to figure out if we want to free the
buffer or not.

Ideally we should save the errno for logger after but this does not
really seem necessary judging by [1] and glibc's relevant
implementation.

Relevant links for further context:
[1]: http://austingroupbugs.net/view.php?id=385
[2]: https://sourceware.org/git/?p=glibc.git;a=blob;f=libio/iogetdelim.c
[3]: https://git.musl-libc.org/cgit/musl/tree/src/stdio/getdelim.c