DEVOPS-1517
Removed static UID/GID, from postinst script. In the pervious release of
ubuntu the UID/GID was not reserverd, in nobel the UID/GID `997` is
being assigned to other services, before the schaufel is created, thus
making the installation fail.

In this PR after discussing the issue with `PE` team, it was decided, to
remove the static UID/GID and let the OS decide, upon installation.

Changed versioning from year number to release codename.
Updated the changelog to bump version up.